feat: integrate GroupService for group management and sorting functionality
