+++ libs/ui/recursive-group-list/src/components/add-group/add-group.component.ts
@@ -1,7 +1,8 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { Component, EventEmitter, inject, Input, Output } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { GroupModel } from '../../models/group-model';
+import { GroupService } from '../../services/group.service';
 
 @Component({
   selector: 'clx-add-group',
@@ -11,6 +12,8 @@ import { GroupModel } from '../../models/group-model';
 export class AddGroupComponent {
   @Input() groups: GroupModel[] = [];
   @Output() groupAdded = new EventEmitter<{ group: GroupModel }>();
+
+  groupService = inject(GroupService);
 
   isModalOpen = false;
   newGroup: GroupModel = {
@@ -40,17 +43,11 @@ export class AddGroupComponent {
     this.newGroup.id = crypto.randomUUID();
     this.groupAdded.emit({ group: this.newGroup });
 
-    this.sortGroupsRecursively(this.groups);
+    const sortedGroups = this.groupService.sortGroupsListRecursively(this.groups);
+    console.log('Sorted groups: ', sortedGroups);
 
     // Reset form and close modal
     this.newGroup = { id: '', name: '', description: '', expanded: false, parentId: undefined, subGroups: [] };
     this.isModalOpen = false;
   }
-
-  private sortGroupsRecursively = (groups: GroupModel[]): GroupModel[] => {
-    return groups.sort((a, b) => a.name.localeCompare(b.name)).map(group => ({
-      ...group,
-      subGroups: group.subGroups ? this.sortGroupsRecursively(group.subGroups) : []
-    }));
-  };
 }
